requirements.txt
Django>=5.0,<6.0
django-environ>=0.11
cloudinary>=1.36
django-cloudinary-storage>=0.3
whitenoise>=6.7
django-ckeditor>=6.7
psycopg2-binary
gunicorn
dj-database-url
django-anymail